reactive-programming-tutorial/src/test/java/com/ncodersalab/reactiveprogramming/services/BookServiceTest.java
package com.ncodersalab.reactiveprogramming.services;

import com.ncodersalab.reactiveprogramming.domain.Book;
import org.junit.jupiter.api.Test;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import reactor.test.StepVerifier;

import static org.junit.jupiter.api.Assertions.assertEquals;

class BookServiceTest {

    private BookInfoService bookInfoService = new BookInfoService();
    private ReviewService reviewService = new ReviewService();
    private BookService bookService = new BookService(bookInfoService, reviewService);

    @Test
    void getBooks() {
        Flux<Book> books = bookService.getBooks();

        StepVerifier.create(books)
                .assertNext(book -> {
                    assertEquals("Book One", book.getBookInfo().getTitle());
                })
                .assertNext(book -> {
                    assertEquals("Book Two", book.getBookInfo().getTitle());
                })
                .assertNext(book -> {
                    assertEquals("Book Three", book.getBookInfo().getTitle());
                }).verifyComplete();

    }

    @Test
    void getBookById() {
        long bookId = 2;
        Mono<Book> book = bookService.getBookById(bookId);
        StepVerifier.create(book)
                .assertNext(b -> {
                    assertEquals("Book " + bookId, b.getBookInfo().getTitle());
                    assertEquals(4, b.getReviews().size());
                }).verifyComplete();
    }
}